feat(smart-accounts): Check for smart account existence and return correct capabilities (#275)
### Description

Part of
[ENG-645](https://linear.app/valora/issue/ENG-645/add-support-for-eip-7702-delegation).
Adds basic support for checking if a user has delegated their EOA to a
Kernel 7702 account, per [this
documentation](https://docs.pimlico.io/guides/eip7702/demo#sending-the-transaction),
behind a feature gate. If the feature gate is enabled, we return wallet
capabilities based off the status of a user's 7702 delegation, instead
of returning the defaults.

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)