chore(walletconnect): tidy up types (#292)
### Description

Since `wallet_getCallsStatus` request has a different property set than
`wallet_getCapabilities`, it is no longer make sense to join them within
`NonInteractiveRequest` type.

Also the `NonInteractiveActions` type and `isNonInteractiveAction` is
deprecated in favor of `requiresUserConsent()` which basically does the
same but has more explicit naming.

### Test plan

CI

### Related issues

- Related to ENG-644

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)