fix(send): Move internal send navigation to before receipt fetching (#286)
### Description

~~Error description / logs in the ticket, it seems like receipt fetching
could fail if a node wasn't up to date after the transaction submitted,
added exponential backoff retry logic (5 attempts max 10 seconds, can
change if we want something different) to give it a little time to get
the receipt before actually failing.~~

Also moved the navigation to before the receipt checking (for internal
sends), to not block the user after submitting the transaction. This
follows the same pattern as is in the earn and swap transaction sagas.

### Test plan

Ran wallet locally and did multiple sends on Celo, all worked as
expected and navigated me back to home screen after. However looking at
the logs all succeeded on their first receipt fetching attempt, and not
sure how to reproduce the bug. But the happy path at least works as
expected.

Tried to add unit test but was having difficulty mocking the
`publicClient.celo` calls, and cursor AI was just going in loops adding
more and more complex code, so just did the manual testing instead.

### Related issues

- Fixes ENG-663

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)